+++ src/modules/plugins/plugins.service.ts
@@ -158,8 +158,9 @@ export class PluginsService {
     const modules = await this.getInstalledModules();
     const disabledPlugins = await this.getDisabledPlugins();
 
-    // filter out non-homebridge plugins by name
+    // filter out the ui and non-homebridge plugins by name
     const homebridgePlugins = modules
+      .filter(module => module.name !== this.configService.name)
       .filter(module => (module.name.indexOf('homebridge-') === 0) || this.isScopedPlugin(module.name))
       .filter(module => pathExistsSync(join(module.installPath, 'package.json')));
 
@@ -169,11 +170,11 @@ export class PluginsService {
     await Promise.all(homebridgePlugins.map(async (pkg) => {
       return limit(async () => {
         try {
-          const pjson: IPackageJson = await readJson(join(pkg.installPath, 'package.json'));
+          const pkgJson: IPackageJson = await readJson(join(pkg.installPath, 'package.json'));
           // check each plugin has the 'homebridge-plugin' keyword
-          if (pjson.keywords && pjson.keywords.includes('homebridge-plugin')) {
+          if (pkgJson.keywords && pkgJson.keywords.includes('homebridge-plugin')) {
             // parse the package.json for each plugin
-            const plugin = await this.parsePackageJson(pjson, pkg.path);
+            const plugin = await this.parsePackageJson(pkgJson, pkg.path);
 
             // check if the plugin has been disabled
             plugin.disabled = disabledPlugins.includes(plugin.name);
@@ -193,15 +194,7 @@ export class PluginsService {
     }));
 
     this.installedPlugins = plugins;
-    return orderBy(
-      plugins,
-      [
-        (resultItem: HomebridgePlugin) => resultItem.name === this.configService.name,
-        'updateAvailable',
-        'name',
-      ],
-      ['desc', 'desc', 'asc'],
-    );
+    return plugins;
   }
 
   /**
@@ -513,10 +506,10 @@ export class PluginsService {
   public async getHomebridgePackage() {
     // try load from the "homebridgePackagePath" option first
     if (this.configService.ui.homebridgePackagePath) {
-      const pjsonPath = join(this.configService.ui.homebridgePackagePath, 'package.json');
-      if (await pathExists(pjsonPath)) {
+      const pkgJsonPath = join(this.configService.ui.homebridgePackagePath, 'package.json');
+      if (await pathExists(pkgJsonPath)) {
         try {
-          return await this.parsePackageJson(await readJson(pjsonPath), this.configService.ui.homebridgePackagePath);
+          return await this.parsePackageJson(await readJson(pkgJsonPath), this.configService.ui.homebridgePackagePath);
         } catch (err) {
           throw err;
         }
@@ -543,8 +536,8 @@ export class PluginsService {
     }
 
     const homebridgeModule = homebridgeInstalls[0];
-    const pjson: IPackageJson = await readJson(join(homebridgeModule.installPath, 'package.json'));
-    const homebridge = await this.parsePackageJson(pjson, homebridgeModule.path);
+    const pkgJson: IPackageJson = await readJson(join(homebridgeModule.installPath, 'package.json'));
+    const homebridge = await this.parsePackageJson(pkgJson, homebridgeModule.path);
 
     if (!homebridge.latestVersion) {
       return homebridge;
@@ -631,8 +624,8 @@ export class PluginsService {
         throw new Error('Could not find npm package');
       }
 
-      const pjson: IPackageJson = await readJson(join(npmPkg.installPath, 'package.json'));
-      const npm = await this.parsePackageJson(pjson, npmPkg.path) as HomebridgePlugin & { showUpdateWarning?: boolean };
+      const pkgJson: IPackageJson = await readJson(join(npmPkg.installPath, 'package.json'));
+      const npm = await this.parsePackageJson(pkgJson, npmPkg.path) as HomebridgePlugin & { showUpdateWarning?: boolean };
 
       // show the update warning if the installed version is below the minimum recommended
       // (bwp91) I set this to 9.5.0 to match a minimum node version of 18.15.0
@@ -1227,31 +1220,31 @@ export class PluginsService {
 
   /**
    * Convert the package.json into a HomebridgePlugin
-   * @param pjson
+   * @param pkgJson
    * @param installPath
    */
-  private async parsePackageJson(pjson: IPackageJson, installPath: string): Promise<HomebridgePlugin> {
+  private async parsePackageJson(pkgJson: IPackageJson, installPath: string): Promise<HomebridgePlugin> {
     const plugin: HomebridgePlugin = {
-      name: pjson.name,
-      private: pjson.private || false,
-      displayName: pjson.displayName,
-      description: (pjson.description) ?
-        pjson.description.replace(/(?:https?|ftp):\/\/[\n\S]+/g, '').trim() : pjson.name,
-      verifiedPlugin: this.verifiedPlugins.includes(pjson.name),
-      icon: this.verifiedPluginsIcons[pjson.name]
-        ? `${this.verifiedPluginsIconsPrefix}${this.verifiedPluginsIcons[pjson.name]}`
+      name: pkgJson.name,
+      private: pkgJson.private || false,
+      displayName: pkgJson.displayName,
+      description: (pkgJson.description) ?
+        pkgJson.description.replace(/(?:https?|ftp):\/\/[\n\S]+/g, '').trim() : pkgJson.name,
+      verifiedPlugin: this.verifiedPlugins.includes(pkgJson.name),
+      icon: this.verifiedPluginsIcons[pkgJson.name]
+        ? `${this.verifiedPluginsIconsPrefix}${this.verifiedPluginsIcons[pkgJson.name]}`
         : null,
-      installedVersion: installPath ? (pjson.version || '0.0.1') : null,
+      installedVersion: installPath ? (pkgJson.version || '0.0.1') : null,
       globalInstall: (installPath !== this.configService.customPluginPath),
-      settingsSchema: await pathExists(resolve(installPath, pjson.name, 'config.schema.json')) || this.miscSchemas[pjson.name],
+      settingsSchema: await pathExists(resolve(installPath, pkgJson.name, 'config.schema.json')) || this.miscSchemas[pkgJson.name],
       installPath,
     };
 
     // only verified plugins can show donation links
-    plugin.funding = plugin.verifiedPlugin ? pjson.funding : undefined;
+    plugin.funding = plugin.verifiedPlugin ? pkgJson.funding : undefined;
 
     // if the plugin is private, do not attempt to query npm
-    if (pjson.private) {
+    if (pkgJson.private) {
       plugin.publicPackage = false;
       plugin.latestVersion = null;
       plugin.updateAvailable = false;